+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # ChangelogChangelog
 
-## 2.0.0a5 (to be released on 31.12.2019)
+## 2.0.0a5 (to be released on 03.01.2020)
 - fixed autocomplete for *.should methods
+- reflected all "old and redundant" SeleneElement methods as deprecated in Element
+  - added corresponding conditions
 
 ## 2.0.0a4 (released on 30.12.2019)
 - fixed default browser_name handling in shared config; implemented some old opts in config